Added exception handler with ResponseEntityExceptionHandler to manage errors and return the appropriate responses

+++ fake-api/src/main/java/com/igor/fake_api/business/service/ProductService.java
@@ -2,6 +2,8 @@ package com.igor.fake_api.business.service;
 
 import com.igor.fake_api.apiv1.dto.ProductDTO;
 import com.igor.fake_api.infrastructure.entities.Product;
+import com.igor.fake_api.infrastructure.exceptions.ProductAlreadyExistsException;
+import com.igor.fake_api.infrastructure.exceptions.ProductUnexistException;
 import com.igor.fake_api.infrastructure.repository.ProductRepository;
 import com.igor.fake_api.utils.ProductUtils;
 import org.springframework.stereotype.Service;
@@ -20,7 +22,7 @@ public class ProductService {
     public void saveProduct(ProductDTO product) {
 
         if (productRepository.existsByTitle(product.getTitle()) ) {
-            throw new RuntimeException("This product is already registered");
+            throw new ProductAlreadyExistsException();
         }
 
         var productEntity = ProductUtils.dtoToEntity(product);
@@ -34,19 +36,29 @@ public class ProductService {
 
     public Product findByTitle(String title) {
 
-        return productRepository.findByTitle(title);
+        var product = productRepository.findByTitle(title);
+
+        if (product == null) {
+            throw new ProductUnexistException();
+        }
+
+        return product;
     }
 
     public void deleteProduct(String title) {
 
         var product = productRepository.findByTitle(title);
+        if (product == null) {
+            throw new ProductUnexistException();
+        }
+
         productRepository.delete(product);
     }
 
     public ProductDTO updateProduct(Long id, ProductDTO request) {
 
         var productEntity = productRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("A product with this ID does not exist"));
+                .orElseThrow(() -> new ProductUnexistException("A product with this ID does not exist"));
 
         var productUpdated = ProductUtils.updateEntity(productEntity, request);
         productRepository.save(productUpdated);
